Moving file response functionality to shared folder (#22226)
* Moving file response functionality to shared folder

* Linter fixes

* Added the attachment serve to shared folder

* Removed the unwanted logs

* import fixes

* Linter fixes

* License added

* Moved UnsafeContentTypes and MediaContentTypes from api4

---------
